# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UniGate API",
    "version": "0.1.0",
    "description": "Vehicle hiring & management platform. Money is a decimal string; timestamps are ISO-8601 UTC; every response uses the same envelope."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "code",
              "requestId"
            ]
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "error"
        ]
      },
...
      "CustomerEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/Customer"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "meta": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "meta"
        ]
      },
...
      "OwnerEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/Owner"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "meta": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "meta"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/customers/{id}": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Profile + corporate extension (own → global)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "customers"
        ],
        "summary": "Default city, VAT number (locked once invoiced), SPO attribution",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "defaultCityId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  },
                  "vatNumber": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "pattern": "^3\\d{13}3$"
                  },
                  "acquiredBySpoId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED with details.lockedByInvoiceNumber",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/owners/{id}": {
      "get": {
        "tags": [
          "owners"
        ],
        "summary": "Owner profile (own → global); nationalIdLast4 only for self or owners.pii.reveal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "owners"
        ],
        "summary": "Business details, national id (encrypted), verticals, privacy; identity edits after approval → UNDER_REVIEW",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ownerType": {
                    "type": "string",
                    "enum": [
                      "INDIVIDUAL",
                      "COMPANY"
                    ]
                  },
                  "businessNameEn": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 160
                  },
                  "businessNameAr": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 160
                  },
                  "crNumber": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "pattern": "^\\d{10}$"
                  },
                  "vatNumber": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "pattern": "^3\\d{13}3$"
                  },
                  "isVatRegistered": {
                    "type": "boolean"
                  },
                  "nationalId": {
                    "type": "string",
                    "pattern": "^[12]\\d{9}$"
                  },
                  "privacySettings": {
                    "type": "object",
                    "properties": {
                      "showBusinessName": {
                        "type": "boolean",
                        "default": true
                      },
                      "showRating": {
                        "type": "boolean",
                        "default": true
                      },
                      "showFleetSize": {
                        "type": "boolean",
                        "default": false
                      },
                      "showCity": {
                        "type": "boolean",
                        "default": true
                      },
                      "showContactPhone": {
                        "type": "boolean",
                        "default": false
                      }
                    },
                    "additionalProperties": false
                  },
                  "transportTypes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "PASSENGER",
                        "GOODS"
                      ]
                    },
                    "minItems": 1,
                    "maxItems": 2
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "OWNER_NOT_APPROVED (suspended)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/drivers/{id}": {
      "get": {
        "tags": [
          "drivers"
        ],
        "summary": "Driver profile — last-4 identifiers only",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "drivers"
        ],
        "summary": "Licence, categories, emergency contact (drivers.update)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idType": {
                    "type": "string",
                    "enum": [
                      "NATIONAL_ID",
                      "IQAMA"
                    ]
                  },
                  "nationalId": {
                    "type": "string",
                    "pattern": "^[12]\\d{9}$"
                  },
                  "dateOfBirth": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "licenseNumber": {
                    "type": "string",
                    "pattern": "^[0-9A-Z]{5,20}$"
                  },
                  "licenseExpiryDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "licenseCategories": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[A-Z0-9_]{1,16}$"
                    },
                    "minItems": 1,
                    "maxItems": 10
                  },
                  "transportTypes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "PASSENGER",
                        "GOODS"
                      ]
                    },
                    "minItems": 1,
                    "maxItems": 2
                  },
                  "emergencyContactName": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 160
                  },
                  "emergencyContactPhone": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "pattern": "^\\+[1-9]\\d{7,14}$"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "drivers"
        ],
        "summary": "Deactivate (blocked on an active trip); sessions revoked",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deactivated"
          },
          "409": {
            "description": "DRIVER_ALREADY_ON_TRIP",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/spo/profiles/{id}": {
      "get": {
        "tags": [
          "spo"
        ],
        "summary": "One SPO",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpoProfileEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "spo"
        ],
        "summary": "Update an SPO (spo.update)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "employeeCode": {
                    "type": "string",
                    "pattern": "^[A-Z0-9-]{2,32}$"
                  },
                  "regionId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  },
                  "managerUserId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  },
                  "commissionModelId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpoProfileEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/spo/profiles/{id}/customers/{customerProfileId}": {
      "delete": {
        "tags": [
          "spo"
        ],
        "summary": "Unassign (spo.update)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "customerProfileId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Unassigned"
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}